reassigned more general ui tests
