+++ cfme/tests/webui/test_advanced_search.py
@@ -5,12 +5,16 @@ import fauxfactory
 import pytest
 import six
 
+from cfme import test_requirements
 from cfme.infrastructure.config_management import ConfigManager
 from cfme.infrastructure.config_management import ConfigSystem
 from cfme.services.myservice import MyService
 from cfme.services.workloads import TemplatesImages
 from cfme.services.workloads import VmsInstances
 from cfme.utils.appliance.implementations.ui import navigate_to
+
+
+pytestmark = [test_requirements.filtering]
 
 Param = namedtuple("Param", ["collection", "destination", "entity", "filter", "my_filters"])
